# use Base.sum instead of handrolled sum with poorly initialized accumulator