--- a/src/multivariates.jl
+++ b/src/multivariates.jl
@@ -113,20 +113,10 @@ end
 
 ## log likelihood
 
-function _loglikelihood(d::MultivariateDistribution, X::AbstractMatrix)
-    ll = 0.0
-    for i in 1:size(X, 2)
-        ll += _logpdf(d, view(X,:,i))
-    end
-    return ll
-end
-
 function loglikelihood(d::MultivariateDistribution, X::AbstractMatrix)
-    size(X,1) == length(d) ||
-        throw(DimensionMismatch("Inconsistent array dimensions."))
-    _loglikelihood(d, X)
+    size(X, 1) == length(d) || throw(DimensionMismatch("Inconsistent array dimensions."))
+    return sum(x -> _logpdf(d, x), X, 2)
 end
-
 
 ##### Specific distributions #####