This results in better type stability in cases where the reduction output is not a `Float64`. I found this while messing about with the [code in this Discourse post](https://discourse.julialang.org/t/julia-static-compilation/296/20) and [a branch of ForwardDiff which adds stack-allocated, low-dimensional gradients](https://github.com/JuliaDiff/ForwardDiff.jl/pull/213).

Note that I found a bunch of places with the same problem throughout the codebase - all of these spots will likely kill performance with non-`Float64` number types like `Dual`.

Setup:

```julia
using BenchmarkTools, ForwardDiff, Distributions

f(x1, x2, D) = loglikelihood(Normal(x1, x2), D)
∇f(x1, x2, D) = ForwardDiff.derivative((y1, y2) -> f(y1, y2, D), (x1, x2))

x1, x2 = rand(2);
D = rand(Normal(5.0, 1.0), 10);
```

Before this PR:

```julia
julia> @benchmark ∇f($x1, $x2, $D)
BenchmarkTools.Trial:
  memory estimate:  1008 bytes
  allocs estimate:  32
  --------------
  minimum time:     704.517 ns (0.00% GC)
  median time:      744.134 ns (0.00% GC)
  mean time:        853.260 ns (12.39% GC)
  maximum time:     20.858 μs (92.76% GC)
```

After this PR:

```julia
julia> @benchmark ∇f($x1, $x2, $D)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     323.714 ns (0.00% GC)
  median time:      323.957 ns (0.00% GC)
  mean time:        325.874 ns (0.00% GC)
  maximum time:     1.221 μs (0.00% GC)
```