@andreasnoack 2017-04-13
It would be great if you could just remove the `_` version. It doesn't really do anything. There is a similar method in `multivariates` which is also uses a literal for initialization of the accumulator so it would be good to have that one fixed as well.